backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
click==8.3.2
croniter==6.2.2
fastapi==0.135.3
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
msgpack==1.1.2
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
py-algorand-sdk==2.11.1
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyNaCl==1.6.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
redis==7.4.0
rq==2.7.0
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
